[operator] Add omatcopy_batch implementation and test
Add omatcopy_batch implementation exploiting the work done for
the omatcopy opearator. In particular, it reuses the single kernel
and passing the correct pointers to the call.

Test:
None reference implementation provides omatcopy_batch operator so
to test it reuses the single omatcopy implementation inside a loop to mimic the
batch.

Signed-off-by: s-Nick <nicolo.scipione@codeplay.com>