Refactor Asum level1 operator
This commit change completely the implementation of the Asum level1
operator. Now it shares the same structure of many level2 operators.
It has different backend files that handle the kernel depending on
following hardware and input size. The kernel code is new and it uses
many sycl2020 features and tries to exploit the sub_group size of each
architecture.

Signed-off-by: nscipione <nicolo.scipione@codeplay.com>