Fix trsm operator using DEFAULT as TUNING_TARGET for NVIDIA GPUs.
Current default implementation uses a different gemm implementation that
hangs on NVIDIA GPUs. In particular, a store function hang for certain
tests. Apparently adding a synchronization point before fixes the hangs.
This is a workaround due to the current situation.
Adding back trsm tests and removing test skipping when this
configuration is used.

Signed-off-by: nscipione <nicolo.scipione@codeplay.com>